No need for the cast
git-svn-id: file:///home/toad/git-migration/temprepository/branches/db4o/freenet@24469 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/client/async/SplitFileFetcherSubSegment.java
@@ -387,7 +387,7 @@ public class SplitFileFetcherSubSegment extends SendableGet implements SupportsB
 			boolean removed = false;
 			synchronized(segment) {
 				for(int i=0;i<blockNums.size();i++) {
-					Integer x = (Integer) blockNums.get(i);
+					Integer x = blockNums.get(i);
 					// Compare by value as sometimes we will do new Integer(num) in requeueing after cooldown code.
 					if(x.equals(token)) {
 						blockNums.remove(i);
